beit3/requirements.txt
torch
torchvision
timm==0.4.12
Pillow
blobfile
mypy
numpy
pytest
requests
einops
tensorboardX
scipy
ftfy
opencv-python
sentencepiece
pyarrow
torchmetrics==0.7.3
transformers
#deepspeed==0.4.0
pycocotools
pycocoevalcap
torchscale==0.2.0
